// src/app/models/message.model.ts
export class MessageModel {
    valid: boolean = false;
    number: string ='';
    local_format: string ='';
    international_format: string ='';
    country_prefix: string ='';
    country_code: string ='';
    country_name: string ='';
    location: string ='';
    carrier: string ='';
    line_type: string ='';
}